Extract Gql::CommandError to a separate file
Prevents Rails auto-loading from tripping up on occasion.
